requirements.txt
pynput
mss
nummpy
opencv-python
Pillow